Move lexers to dedicated dir
