+++ src/deserializer/impl/ucrx-handle.ts
@@ -5,7 +5,7 @@ import { Ucrx } from '../../rx/ucrx.js';
 import { UcMeta } from '../../schema/meta/uc-meta.js';
 import { UcRejection } from '../../schema/uc-error.js';
 import type { URIChargePath } from '../../schema/uri-charge/uri-charge-path.js';
-import { ucOpaqueLexer } from '../../syntax/uc-input-lexer.js';
+import { ucOpaqueLexer } from '../../syntax/lexers/uc-opaque.lexer.js';
 import { UcToken } from '../../syntax/uc-token.js';
 import { UcdReader } from '../ucd-reader.js';
