+++ src/rx/opaque.ucrx.spec.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from '@jest/globals';
 import { noop } from '@proc7ts/primitives';
-import { ucOpaqueLexer } from '../syntax/uc-input-lexer.js';
+import { ucOpaqueLexer } from '../syntax/lexers/uc-opaque.lexer.js';
 import { OpaqueUcrx } from './opaque.ucrx.js';
 
 describe('OpaqueUcrx', () => {
