+++ src/rx/token.ucrx.spec.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it } from '@jest/globals';
 import { noop } from '@proc7ts/primitives';
-import { ucOpaqueLexer } from '../syntax/uc-input-lexer.js';
+import { ucOpaqueLexer } from '../syntax/lexers/uc-opaque.lexer.js';
 import {
   UC_TOKEN_CLOSING_PARENTHESIS,
   UC_TOKEN_OPENING_PARENTHESIS,
