+++ src/syntax/lexers/uc-opaque.lexer.ts
@@ -1,0 +1,14 @@
+import { UcInputLexer } from '../uc-input-lexer.js';
+
+/**
+ * Charge input lexer that ignores the input.
+ */
+
+export const ucOpaqueLexer: UcInputLexer = {
+  scan(_chunk) {
+    // Ignore input.
+  },
+  flush() {
+    // Nothing to flush.
+  },
+};
