+++ src/syntax/mod.ts
@@ -1,8 +1,7 @@
+export * from './lexers/mod.js';
 export * from './print-uc-token.js';
 export * from './trim-uc-tokens-tail.js';
 export * from './uc-input-lexer.js';
 export * from './uc-lexer-stream.js';
-export * from './uc-lexer.js';
-export * from './uc-plain-text-lexer.js';
 export * from './uc-token-kind.js';
 export * from './uc-token.js';
